src/requirements.txt
gunicorn==20.0.4
Django==2.2.4
django-debug-toolbar==2.0
django-dotenv==1.4.2
django-extensions==2.2.1
Pillow==6.1.0
psycopg2-binary
Werkzeug==0.15.5
djangorestframework-datatables==0.5.0
simplejson==3.17.0
azure-servicebus==0.50.2
celery==4.3.0
redis==3.3.8
django-regex-field==1.0.0
-e git+https://github.com/saxix/django-regex.git@a9b3c243e996139251957ea390ae7588aad033c6#egg=django-regex
django-storages==1.9.1
azure-storage-blob==2.1.0
json_log_formatter==0.3.0
beautifulsoup4==4.9.0
-e git+https://github.com/jobec/django-auth-adfs.git@41fa2b8eca8f80be9db2026a6310673f9cbde1bb#egg=django-auth-adfs
azure-identity==1.3.0
azure-keyvault-secrets==4.1.0
Wand==0.6.1
pytesseract==0.3.4
Pillow==6.1.0